# fix(project): config merging