--- a/crates/biome_cli/src/commands/check.rs
+++ b/crates/biome_cli/src/commands/check.rs
@@ -10,8 +10,7 @@ use biome_service::configuration::{
     load_configuration, FormatterConfiguration, LinterConfiguration, LoadedConfiguration,
 };
 use biome_service::workspace::{FixFileMode, UpdateSettingsParams};
-use biome_service::{Configuration, MergeWith};
-use biome_service::{Configuration, ConfigurationBasePath};
+use biome_service::Configuration;
 use std::ffi::OsString;
 
 pub(crate) struct CheckCommandPayload {