--- a/crates/biome_cli/src/commands/format.rs
+++ b/crates/biome_cli/src/commands/format.rs
@@ -7,15 +7,18 @@ use crate::{
     execute_mode, setup_cli_subscriber, CliDiagnostic, CliSession, Execution, TraversalMode,
 };
 use biome_console::{markup, ConsoleExt};
+use biome_deserialize::{MergeWith, NoneState};
 use biome_diagnostics::PrintDiagnostic;
 use biome_service::configuration::css::CssFormatter;
 use biome_service::configuration::json::JsonFormatter;
 use biome_service::configuration::vcs::VcsConfiguration;
 use biome_service::configuration::{
-    load_configuration, FilesConfiguration, FormatterConfiguration, LoadedConfiguration,
+    load_configuration, CssConfiguration, FilesConfiguration, FormatterConfiguration,
+    JavascriptConfiguration, JsonConfiguration, LoadedConfiguration,
 };
 use biome_service::workspace::UpdateSettingsParams;
 use biome_service::{JavascriptFormatter, MergeWith};
+use biome_service::{Configuration, ConfigurationBasePath, JavascriptFormatter};
 use std::ffi::OsString;
 
 pub(crate) struct FormatCommandPayload {
@@ -39,16 +42,16 @@ pub(crate) fn format(
     payload: FormatCommandPayload,
 ) -> Result<(), CliDiagnostic> {
     let FormatCommandPayload {
-        javascript_formatter,
-        formatter_configuration,
+        mut javascript_formatter,
+        mut formatter_configuration,
         vcs_configuration,
         mut paths,
         cli_options,
         stdin_file_path,
         files_configuration,
         write,
-        json_formatter,
-        css_formatter,
+        mut json_formatter,
+        mut css_formatter,
         since,
         changed,
     } = payload;
@@ -67,64 +70,91 @@ pub(crate) fn format(
         ..
     } = loaded_configuration;
     // TODO: remove in biome 2.0
-    if formatter_configuration
-        .as_ref()
-        .is_some_and(|f| f.indent_size.is_some())
-    {
-        let console = &mut session.app.console;
-        let diagnostic = DeprecatedArgument::new(markup! {
-            "The argument "<Emphasis>"--indent-size"</Emphasis>" is deprecated, it will be removed in the next major release. Use "<Emphasis>"--indent-width"</Emphasis>" instead."
-        });
-        console.error(markup! {
-            {PrintDiagnostic::simple(&diagnostic)}
-        })
+    if let Some(config) = formatter_configuration.as_mut() {
+        if let Some(indent_size) = config.indent_size {
+            let console = &mut session.app.console;
+            let diagnostic = DeprecatedArgument::new(markup! {
+                "The argument "<Emphasis>"--indent-size"</Emphasis>" is deprecated, it will be removed in the next major release. Use "<Emphasis>"--indent-width"</Emphasis>" instead."
+            });
+            console.error(markup! {
+                {PrintDiagnostic::simple(&diagnostic)}
+            });
+
+            config.indent_width = Some(indent_size);
+        }
     }
     // TODO: remove in biome 2.0
-    if javascript_formatter
-        .as_ref()
-        .is_some_and(|f| f.indent_size.is_some())
-    {
-        let console = &mut session.app.console;
-        let diagnostic = DeprecatedArgument::new(markup! {
-            "The argument "<Emphasis>"--javascript-formatter-indent-size"</Emphasis>" is deprecated, it will be removed in the next major release. Use "<Emphasis>"--javascript-formatter-indent-width"</Emphasis>" instead."
-        });
-        console.error(markup! {
-            {PrintDiagnostic::simple(&diagnostic)}
-        })
+    if let Some(js_formatter) = javascript_formatter.as_mut() {
+        if let Some(indent_size) = js_formatter.indent_size {
+            let console = &mut session.app.console;
+            let diagnostic = DeprecatedArgument::new(markup! {
+                "The argument "<Emphasis>"--javascript-formatter-indent-size"</Emphasis>" is deprecated, it will be removed in the next major release. Use "<Emphasis>"--javascript-formatter-indent-width"</Emphasis>" instead."
+            });
+            console.error(markup! {
+                {PrintDiagnostic::simple(&diagnostic)}
+            });
+
+            js_formatter.indent_width = Some(indent_size);
+        }
     }
     // TODO: remove in biome 2.0
-    if json_formatter
-        .as_ref()
-        .is_some_and(|f| f.indent_size.is_some())
-    {
-        let console = &mut session.app.console;
-        let diagnostic = DeprecatedArgument::new(markup! {
-            "The argument "<Emphasis>"--json-formatter-indent-size"</Emphasis>" is deprecated, it will be removed in the next major release. Use "<Emphasis>"--json-formatter-indent-width"</Emphasis>" instead."
-        });
-        console.error(markup! {
-            {PrintDiagnostic::simple(&diagnostic)}
-        })
+    if let Some(json_formatter) = json_formatter.as_mut() {
+        if let Some(indent_size) = json_formatter.indent_size {
+            let console = &mut session.app.console;
+            let diagnostic = DeprecatedArgument::new(markup! {
+                "The argument "<Emphasis>"--json-formatter-indent-size"</Emphasis>" is deprecated, it will be removed in the next major release. Use "<Emphasis>"--json-formatter-indent-width"</Emphasis>" instead."
+            });
+            console.error(markup! {
+                {PrintDiagnostic::simple(&diagnostic)}
+            });
+
+            json_formatter.indent_width = Some(indent_size);
+        }
     }
     // TODO: remove in biome 2.0
-    if css_formatter
-        .as_ref()
-        .is_some_and(|f| f.indent_size.is_some())
-    {
-        let console = &mut session.app.console;
-        let diagnostic = DeprecatedArgument::new(markup! {
-            "The argument "<Emphasis>"--css-formatter-indent-size"</Emphasis>" is deprecated, it will be removed in the next major release. Use "<Emphasis>"--css-formatter-indent-width"</Emphasis>" instead."
-        });
-        console.error(markup! {
-            {PrintDiagnostic::simple(&diagnostic)}
-        })
+    if let Some(css_formatter) = css_formatter.as_mut() {
+        if let Some(indent_size) = css_formatter.indent_size {
+            let console = &mut session.app.console;
+            let diagnostic = DeprecatedArgument::new(markup! {
+                "The argument "<Emphasis>"--css-formatter-indent-size"</Emphasis>" is deprecated, it will be removed in the next major release. Use "<Emphasis>"--css-formatter-indent-width"</Emphasis>" instead."
+            });
+            console.error(markup! {
+                {PrintDiagnostic::simple(&diagnostic)}
+            });
+
+            css_formatter.indent_width = Some(indent_size);
+        }
     }
 
-    configuration.merge_with(javascript_formatter);
-    configuration.merge_with(json_formatter);
-    configuration.merge_with(css_formatter);
-    configuration.merge_with(formatter_configuration);
-    configuration.merge_with(vcs_configuration);
-    configuration.merge_with(files_configuration);
+    configuration.merge_with(Configuration {
+        css: Some(CssConfiguration {
+            formatter: css_formatter,
+            ..NoneState::none()
+        }),
+        files: files_configuration,
+        formatter: if configuration
+            .formatter
+            .as_ref()
+            .is_some_and(FormatterConfiguration::is_disabled)
+        {
+            None
+        } else {
+            Some(FormatterConfiguration {
+                enabled: Some(true),
+                ..formatter_configuration.unwrap_or_else(NoneState::none)
+            })
+        },
+        javascript: Some(JavascriptConfiguration {
+            formatter: javascript_formatter,
+            ..NoneState::none()
+        }),
+        json: Some(JsonConfiguration {
+            formatter: json_formatter,
+            ..NoneState::none()
+        }),
+        vcs: vcs_configuration,
+        ..NoneState::none()
+    });
 
     // check if support of git ignore files is enabled
     let vcs_base_path = configuration_path.or(session.app.fs.working_directory());