--- a/crates/biome_deserialize/src/lib.rs
+++ b/crates/biome_deserialize/src/lib.rs
@@ -32,11 +32,16 @@
 mod diagnostics;
 mod impls;
 pub mod json;
+mod merge;
+mod none_state;
 pub mod string_set;
+
 use biome_diagnostics::{Error, Severity};
 use biome_rowan::TextRange;
 pub use diagnostics::{DeserializationAdvice, DeserializationDiagnostic, VisitableType};
 pub use impls::*;
+pub use merge::MergeWith;
+pub use none_state::NoneState;
 use std::fmt::Debug;
 pub use string_set::StringSet;