--- a/crates/biome_deserialize/src/none_state.rs
+++ b/crates/biome_deserialize/src/none_state.rs
@@ -1,0 +1,7 @@
+pub trait NoneState {
+    /// Returns the "none state" for a deserializable type.
+    ///
+    /// This represents the state in which no fields have been set and the
+    /// defaults have not yet been filled in. I.e. every field is `None`.
+    fn none() -> Self;
+}