… CHANGELOG.md
… Cargo.lock
… Cargo.toml
… crates/biome_cli/src/lib.rs
… crates/biome_cli/tests/cases/config_extends.rs
… crates/biome_cli/tests/snapshots/main_cases_config_extends/allows_reverting_fields_in_extended_config_to_default.snap
… crates/biome_cli/tests/snapshots/main_cases_config_extends/extends_config_ok_linter_not_formatter.snap
… crates/biome_cli/tests/snapshots/main_cases_config_extends/extends_config_ok_linter_not_formatter.snap.new
… crates/biome_cli/tests/snapshots/main_cases_config_extends/respects_unaffected_values_from_extended_config.snap
… crates/biome_cli/tests/snapshots/main_commands_check/should_not_enable_all_recommended_rules.snap.new
… crates/biome_cli/tests/snapshots/main_commands_ci/correctly_handles_ignored_and_not_ignored_files.snap.new
… crates/biome_deserialize/Cargo.toml
… crates/biome_deserialize_macros/Cargo.toml
… crates/biome_deserialize_macros/src/lib.rs
… crates/biome_deserialize_macros/src/merge_derive.rs
… crates/biome_deserialize_macros/src/mergeable_derive.rs
… crates/biome_deserialize_macros/src/none_state_derive.rs
… crates/biome_formatter/Cargo.toml
… crates/biome_formatter/src/lib.rs
… crates/biome_js_formatter/Cargo.toml
… crates/biome_js_formatter/src/context.rs
… crates/biome_js_formatter/src/context/trailing_comma.rs
… crates/biome_service/Cargo.toml
… crates/biome_service/src/configuration/css.rs
… crates/biome_service/src/configuration/formatter.rs
… crates/biome_service/src/configuration/javascript/formatter.rs
… crates/biome_service/src/configuration/javascript/mod.rs
… crates/biome_service/src/configuration/json.rs
… crates/biome_service/src/configuration/linter/mod.rs
… crates/biome_service/src/configuration/linter/rules.rs