## Summary

This is my first shot at fixing #1440. There are two main changes:

- All the configuration structs (the ones where all the fields are already wrapped in `Option`) now have a `NoneState` in addition to the `Default` state. As you might guess, the `NoneState` represents the state where all the `Option`s are `None`. This state is used as the starting point for deserialization instead of `Default`, which allows merging to detect which fields may be overridden, and which ones shouldn't.
- The `MergeWith` trait has been replaced with a `Merge` trait, which has also been moved from `biome_service` to `biome_deserialize`. All the implementations of `MergeWith` have been removed in favor of a `Mergeable` derive macro that generates all the `Merge` implementations.

It is worth mentioning that the `Merge` trait has a dedicated `merge_in_defaults()` method, which we use to merge in the defaults after merging of `extends` configs has been done. The reason it requires a separate method (as opposed to simply `Configuration::default().merge_with(config)`) is because `merge_with()` overrides `None` fields with the value from the struct it is merging with, while in the case of merging defaults we need to also recursively check if that value has other `None` fields for which a `Default` might be available.

I have added test cases for the issue reported in #1440 and another to highlight that it is possible for a config that extends another to revert values to their default value (which the original `merge_if_not_default()` method would have had trouble with).

One complication I ran into is that the configurations coming out of `bpaf` arguments are initialized with `Default` instead of `NoneState`, and yet we want to merge them into the configuration from the config file. This was previously resolved by custom `MergeWith` implementations, although I have to say this was easy to overlook. I have now resolved this by explicitly patching the configs as needed in the `check`/`lint`/`format` commands. It's not super pretty, but at least it's explicit and localized.

## Test Plan

I only had to update 1 test case (see comment) and added two more test cases to demonstrate more merging behavior. Still, I might very well have missed something (especially around the CLI invocations), so it would be great if someone could help identify pain points.

## Future Work

I think it is possible to simplify the `Override` structs as well based on this, but I've left it out of scope for now since the PR is big enough as it is.

There is also still an issue where rule options cannot be deeply merged. I've left it for now, since I don't think it's a regression. I expect it's also not very high priority, but it would be nice if it would work as expected. It will require wrapping all the rule options in `Option` as well though.