@Conaclos 2024-01-07
> The `MergeWith` trait has been replaced with a `Merge` trait, which has also been moved from `biome_service` to `biome_deserialize`. 

I am not sure that `biome_deserialize` is the good place for this trait and some of its impls...

> All the configuration structs (the ones where all the fields are already wrapped in Option) now have a NoneState in addition to the Default state.
> One complication I ran into is that the configurations coming out of bpaf arguments are initialized with Default instead of NoneState

Couldn't this be simpler to remove the current implementation of `Default` and use `derive(Default)` that set to `None` all optional fields?

Instead of merging the deserialized and overridden state with a default state, we could expose queries.
Let's me take an example:

```rs
#[derive(Default, Mergeable)]
struct Config {
  enabled: Option<bool>,
  threshold: Option<u32>,
}

impl Config {
  fn enabled(&self) -> bool {
    self.enabled.unwrap_or(true)
  }

  fn threshold(&self) -> u32 {
    self.threshold.unwrap_or(255)
  }
}

fn main() {
  let first_config = Config::default();
  let override_config = Config { enabled: Some(false) };
  let final = first_config.merge_with(override_config);
  assert_eq!(final.enabled(), false);
  assert_eq!(final.threshold(), 255);
}
```