@ematipico 2024-01-08
I haven't seen any issue in terms of performance:

PR
```
  Time (mean ± σ):     825.5 ms ±  15.7 ms    [User: 141.4 ms, System: 72.3 ms]
  Range (min … max):   801.5 ms … 855.9 ms    10 runs
```

`main`
```
  Time (mean ± σ):     821.1 ms ±   9.8 ms    [User: 141.2 ms, System: 66.5 ms]
  Range (min … max):   806.5 ms … 840.8 ms    10 runs
```