@ematipico 2024-01-08
Personally, I think this solution works.

Sure, `Merge` doesn't feel that it belongs to the deserializer crate, but I don't mind that much, considering the issue we're trying to fix.