@Conaclos 2024-01-09
>  Just like StringSet is part of biome_deserialize because it's a common type we deserialize to, so too do we commonly want our deserialized types to be mergeable.

I still think it is not the good place for these traits. We could move them (also `StringSet`) in a dedicated crate (`biome_configuration` or something else). However, I will not block the PR for this. We could do that later.

By the way, I am not fond of the `NoneState` trait and the `Merge::merge_in_defaults` method. However, I will not block the PR for that.