@ematipico 2024-01-09
> By the way, I am not fond of the NoneState trait and the Merge::merge_in_defaults method. However, I will not block the PR for that.

What's your opinion on this? It would be great to collect some feedback on the PR and then address it later.