@Conaclos 2024-01-09
> > By the way, I am not fond of the NoneState trait and the Merge::merge_in_defaults method. However, I will not block the PR for that.
> 
> What's your opinion on this? It would be great to collect some feedback on the PR and then address it later.

`NoneState::none` is only used inside the deserializers of objects and maps. This seems like overkill to introduce a trait just for that. Instead of introducing the trait `NoneState`, we could directly set the fields of a struct with the defaulkt value of the field type (that is `None` in the case of `Option`.
Moreover, in the future, we could automatically generate most of deserializers. I am not comfortable with the idea of requiring that every `Deserializable` type to also implement `NoneState`.

Merge::merge_in_defaults looks like a code sniff to me. However, I don't know exactly what's wrong with it - it's more of an intuition. I will comment here if anything comes to mind.