@arendjr 2024-01-09
To be honest I was a bit bumped I needed `merge_in_defaults()` as well, but it’s necessary because unlike merging in an instance, recursion shouldn’t stop at `None` values.

For instance, `Configuration::formatter` has a default value of `None`, so if we merged in `Configuration::default()` like any other instance, no fields of `FormatterConfiguration` would be initialized. `merge_in_defaults()` makes sure recursion can continue by instantiating  `FormatterConfiguration::default()` when needed.

Short of using queries, like you suggested, I don’t see how we can avoid that.

I do like the idea of generating all deserializers, btw. Maybe once we do that, we could just let the macro for it generate a `Deserializable::none()` method instead of needing it as a separate trait indeed?