@Conaclos 2024-01-09
> I do like the idea of generating all deserializers, btw. Maybe once we do that, we could just let the macro for it generate a Deserializable::none() method instead of needing it as a separate trait indeed?

We should just generate the necessary code directly in the deserializer, no need of introducing a `none` method.

> To be honest I was a bit bumped I needed merge_in_defaults() as well, but it’s necessary because unlike merging in an instance, recursion shouldn’t stop at None values.

Yes, I understand your point.