@arendjr 2024-01-09
@Conaclos I think I may also have another proposal that hopefully addresses most of your concerns. 

What if we just got rid of all the `Option` wrappers in `Configuration` and its related structs? Then we could make a new derive macro, `Partial`, that generates a second struct in which all the fields are automatically wrapped in `Option` again. That secondary struct would derive `Default` and `Merge`, and we would also generate a `From` impl for it. Like this:

```rs
#[derive(Partial)]
struct Configuration {
    formatter: FormatterConfiguration
}
```

Would generate:

```rs
#[derive(Default, Deserializable, Merge)]
struct PartialConfiguration {
    formatter: Option<PartialFormatterConfiguration>,
}

impl From<PartialConfiguration> for Configuration {
    fn from(other: PartialConfiguration) -> Self {
        Self {
            formatter: other
                .formatter
                .map(FormatterConfiguration::from)
                .unwrap_or_default()
        }
    }
}
```

The `From` impl effectively replaces the `merge_in_defaults()` method and because the Partial struct has all the `None` values, which we force by generating it, the regular struct remains free to specify semantic defaults like we do today, and no separate `NoneState` trait is needed.

What I personally really like about this is that it makes it very easy to extend mergeability to the lint rule options, which currently all lack the `Option` wrappers.

But there’s two difficulties with this approach still:
* How does the `Partial` derive on `Configuration` know that `FormatterConfiguration` also had a `Partial` derive? It cannot assume that for all fields.
* `Deserializable` would be implemented on the Partial struct. We *could* still use the handwritten impls we have today, but it feels wrong to hand-maintain impls for generated structs, so we’re practically making generation of the deserializers a blocker for this as well.

What do you think?