@Conaclos 2024-01-10
That's an interesting approach. I also thought about something like this (although less achieved than yours).

However, I am not comfortable of generating a `Partial` that automatically derives `Deserializable`, `serde::{Deserialize, Serialize}`, `schemars:JsonSchema`, and `Bpaf`. This breaks modularity and composability.
Unless we use a declarative macro-rule?

```rs
create_partial! {
  #[derive(Bpaf, Deserializable, Deserialize, ...)}
  struct MyConfiguration {
    prop: bool
  }
}
```

This also implies many code change?

Feel free to open another PR to propose a derive macro for `Deserializable`.

I have to think about this...